requirements.txt
numpy>=1.23.0
opencv-python>=4.8.0
Pillow>=9.0.0
PyYAML>=6.0
requests>=2.31.0
scipy>=1.10.0
tqdm>=4.66.0
psutil>=5.9.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
thop>=0.1.1.post2209072238
